feat(integrations): add tabs for configurations page
