+++ gravitee-apim-console-webui/src/management/integrations/components/integration-status/integration-status.component.html
