+++ gravitee-apim-console-webui/src/management/integrations/integration-configuration/integration-configuration.component.scss
@@ -31,12 +31,6 @@ $typography: map.get(gio.$mat-theme, typography);
   gap: 16px;
 }
 
-.page-cards {
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-}
-
 .page-header {
   &__page-title {
     @include mat.typography-level($typography, headline-6);
@@ -45,64 +39,6 @@ $typography: map.get(gio.$mat-theme, typography);
 
   &__description {
     color: mat.get-color-from-palette(gio.$mat-space-palette, 'lighter40');
-  }
-}
-
-.card-header {
-  padding: 16px;
-  border-bottom: 1px solid mat.get-color-from-palette(gio.$mat-dove-palette, 'darker20');
-
-  &__title {
-    h3 {
-      @include mat.typography-level($typography, subtitle-1);
-      margin: 0;
-    }
-  }
-}
-
-.form {
-  grid-column-start: 2;
-  margin: 16px 0;
-
-  .form-field {
-    width: 100%;
-  }
-
-  textarea {
-    resize: none;
-  }
-
-  .hint {
-    color: mat.get-color-from-palette(gio.$mat-space-palette, 'lighter40');
-  }
-}
-
-.body-strong {
-  @include mat.typography-level($typography, body-1);
-}
-
-.loadingTemplate {
-  height: 50vh;
-  display: flex;
-}
-
-.danger-card {
-  border: 1px solid mat.get-color-from-palette(gio.$mat-error-palette, 'default');
-
-  &__title {
-    color: mat.get-color-from-palette(gio.$mat-error-palette, 'default');
-  }
-
-  &__actions {
-    display: flex;
-    flex-direction: column;
-
-    .danger-action {
-      display: flex;
-      align-items: center;
-      flex-direction: row;
-      justify-content: space-between;
-      height: 48px;
-    }
+    @include mat.typography-level($typography, body-1);
   }
 }
