+++ gravitee-apim-console-webui/src/management/integrations/integrations-navigation/integrations-navigation.component.html
@@ -28,7 +28,12 @@
         </div>
 
         <ng-container *ngFor="let item of allowedItems">
-          <a [routerLink]="item.routerLink" routerLinkActive #rla="routerLinkActive" [routerLinkActiveOptions]="{ exact: true }">
+          <a
+            [routerLink]="item.routerLink"
+            routerLinkActive
+            #rla="routerLinkActive"
+            [routerLinkActiveOptions]="item?.routerLinkActiveOptions ?? { exact: false }"
+          >
             <gio-submenu-item tabIndex="1" [active]="rla.isActive">
               <div class="menu-item">
                 <mat-icon class="menu-item__icon" [svgIcon]="'gio:' + item.icon"> </mat-icon>
